Adding error messages for umimplemented routines.
Prior to this commit, multiple routines in "rocm_blas.cc" and "rocm_fft.cc"
which did not have an underlying implementation, returned without issuing any
error message.

Because of this testcases that invoke routines will fail down the line (from
the call to these routines) with cryptic errors, which in turn make the root
cause analysis time consuming.

Adding the error messages leads to quicker diagnosis.