Workaround for the bug in __longlong_as_double (TENSORFLOW_USE_ROCM)
Native implementations do not exist for the `double` versions of
1. atomicAdd
2. atomicSub
3. atomicMul
4. atomicDiv

So TF code has it's own implementation for the same. This implementation
calls the `__longlong_as_double` routine. We discovered a bug in the
implementation of that routine on the ROCm platform which was leading to
incorrect results. While that bug is being looked at / fixed, this change
updates the TF code to avoid calling that routine in the ROCm flow.