Enabling bias and maxpool ops on the GPU (for fp16)
It seems that these ops were disabled on the GPU to make the debug build work.
However that breaks a couple of unit testcases that specifically expect these
ops to be run on the GPU for fp16. So enabling them back.

Will need to investigate why enabling them breaks the debug build and fix the
root cause.