Updating the split_v_kernel so that formal/actual datatypes match for all args
Prior to this change, for `split_v_kernel`,
the formal dataype for `total_rows` and `total_cols` arguments was `IntType`, and
the actual datatype in all instances (i.e. calls) for the same args is `int`.

For the case where `IntType` resolves to `long long`, this was creates an implicit
cast invovling a size difference (4 vs 8) on the host--devcie boundary.
This seems to trigger a bug in the ROCm flow, in which the size of args in the call
(i.e. host side) does not match the size of args expected on the device side, which
then leads to other errors.

Fixing this issue by updating the formal datatype for `total_rows` and `total_cols`
to be `int` in the `split_v_kernel`.