Re-enabling support for the existing FP16 implementation in HIP.
The previous commit on this branch deleted support for the existing FP16
implementation in HIP, and added support for the new/upcoming FP16 implementation
in HIP.

This commit adds back in the support for existing FP16 implementation in HIP

The code will now switch between using the existing/new FP16 HIP implementations
based on the HIP_VERSIO_PATCH value (18225 being the threshold value)

This commit does NOT re-use the support for existing FP16 implementation that
was deleted in the previous commit. This version is far less obstrusive, and it
does not require changes on the TF side of the fence, and very little changes on
the Eigen side.